Move NotifcationContentDetail in the content.detail package
