+++ src/main/java/com/blackducksoftware/integration/hub/notification/content/VulnerabilityNotificationContent.java
@@ -27,6 +27,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 import com.blackducksoftware.integration.hub.api.component.AffectedProjectVersion;
+import com.blackducksoftware.integration.hub.notification.content.detail.NotificationContentDetail;
 import com.blackducksoftware.integration.hub.service.model.ProjectVersionDescription;
 
 public class VulnerabilityNotificationContent extends NotificationContent {
